1e4a9f3dd5bba6fae9021a50a45af29ec62dcd1e

+++ ejaf_maintenance_equipment/models/res_config_settings.py
@@ -19,7 +19,7 @@ class ResConfigSettings(models.TransientModel):
         res = super(ResConfigSettings, self).get_values()
         params = self.env['ir.config_parameter'].sudo()
         res.update(
-            reservation_liters=int(params.get_param('reservation_liters'))
+            reservation_liters=float(params.get_param('reservation_liters'))
         )
         return res
